Fix: 댓글관련 유틸함수 분리
